Time: 12 ms (33.89%), Space: 8.2 MB (48.19%) - LeetHub

+++ 0009-palindrome-number/0009-palindrome-number.cpp
@@ -5,7 +5,7 @@ public:
         if(x<0)
             return false;
         
-        long long rem,rev=0;
+        long rem,rev=0;
         while(x!=0)
         {
             rem=x%10;
